requirements.txt
keyboard
Pillow
pyperclip
pystray
screeninfo
